Added Initial Game Files - WIP
Game is currently in a working state, the player can travel around the map with buttons or with arrow keys, the bear also moves around and can catch & kill you.

You cannot move outside of the map, if you try, nothing happens and the bear also stays within the map.